// src/services/central-api/generated/model/tenant-search-output.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PL Cental API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TenantSearchOutput
 */
export interface TenantSearchOutput {
    /**
     * 
     * @type {string}
     * @memberof TenantSearchOutput
     */
    'code'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TenantSearchOutput
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TenantSearchOutput
     */
    'taxId'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TenantSearchOutput
     */
    'status'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TenantSearchOutput
     */
    'createBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TenantSearchOutput
     */
    'createTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TenantSearchOutput
     */
    'updateBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TenantSearchOutput
     */
    'updateTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TenantSearchOutput
     */
    'deleteBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TenantSearchOutput
     */
    'deleteTimestamp'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof TenantSearchOutput
     */
    'isDelete'?: boolean | null;
}